Refactor next.config.mjs to use withNextIntl HOC
