+++ messages/fr.json
@@ -1,0 +1,5 @@
+{
+	"Index": {
+	  "title": "Yo la team!"
+	}
+  }
