+++ next.config.mjs
@@ -1,3 +1,8 @@
+import createNextIntlPlugin from 'next-intl/plugin';
+
+const withNextIntl = createNextIntlPlugin();
+
+
 /** @type {import('next').NextConfig} */
 const nextConfig = {
 	images: {
@@ -9,4 +14,4 @@ const nextConfig = {
 	},
 };
 
-export default nextConfig;
+export default withNextIntl(nextConfig);
